refactor(rate-limits): split keychain.rs by platform
